Move updating cache to a separate subfolder
