+++ posthog/queries/funnels/test/test_funnel.py
@@ -5,6 +5,7 @@ from unittest.mock import patch
 from freezegun import freeze_time
 from rest_framework.exceptions import ValidationError
 
+from posthog.caching.update_caching import update_cache_item
 from posthog.constants import FILTER_TEST_ACCOUNTS, INSIGHT_FUNNELS
 from posthog.models import Action, ActionStep, Element
 from posthog.models.cohort import Cohort
@@ -13,7 +14,6 @@ from posthog.models.instance_setting import get_instance_setting
 from posthog.queries.funnels import ClickhouseFunnel, ClickhouseFunnelActors
 from posthog.queries.funnels.test.breakdown_cases import assert_funnel_results_equal, funnel_breakdown_test_factory
 from posthog.queries.funnels.test.conversion_time_cases import funnel_conversion_time_test_factory
-from posthog.tasks.update_cache import update_cache_item
 from posthog.test.base import (
     APIBaseTest,
     ClickhouseTestMixin,
